Restructure, put everything into lib/
