+++ trickler/peripheral/lib/and-fxfz.js
