+++ trickler/peripheral/lib/auto-mode-characteristic.js
