+++ trickler/peripheral/lib/device-info-service.js
