+++ trickler/peripheral/lib/model-number-characteristic.js
