+++ trickler/peripheral/lib/motor.js
