+++ trickler/peripheral/lib/serial-number-characteristic.js
