+++ trickler/peripheral/lib/stability-characteristic.js
