+++ trickler/peripheral/lib/target-weight-characteristic.js
