+++ trickler/peripheral/lib/trickler-service.js
