+++ trickler/peripheral/lib/trickler.js
