+++ trickler/peripheral/lib/unit-characteristic.js
