+++ trickler/peripheral/lib/weight-characteristic.js
